requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.28.0
sse-starlette==2.1.3

# LangGraph / LangChain
langgraph==0.2.57
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.10

# Embeddings
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.0

# Vector store
psycopg2-binary==2.9.10
pgvector==0.3.5
sqlalchemy==2.0.36

# PDF / doc processing
pypdf==5.1.0
python-docx==1.1.2

# RAGAS evals
ragas==0.2.6
datasets==3.2.0

# MLflow
mlflow==2.18.0

# Prometheus
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
tenacity==9.0.0
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1

# Cross-encoder reranking is provided by sentence-transformers above
